feat: remove proposal legacy
