+++ frontend/src/lib/components/proposal-detail/IneligibleNeuronsCard.svelte
@@ -2,10 +2,7 @@
   import { ineligibleNeurons as filterIneligibleNeurons } from "@dfinity/nns";
   import type { ProposalInfo, NeuronInfo } from "@dfinity/nns";
   import { i18n } from "../../stores/i18n";
-  import CardInfo from "../ui/CardInfo.svelte";
-  import type { SvelteComponent } from "svelte";
-  import { VOTING_UI } from "../../constants/environment.constants";
-  import ContentCell from "../ui/ContentCell.svelte";
+  import ProposalContentCell from "./ProposalContentCell.svelte";
 
   export let proposalInfo: ProposalInfo;
   export let neurons: NeuronInfo[];
@@ -23,14 +20,10 @@
     createdTimestampSeconds > proposalInfo.proposalTimestampSeconds
       ? $i18n.proposal_detail__ineligible.reason_since
       : $i18n.proposal_detail__ineligible.reason_short;
-
-  // TODO(L2-965): delete legacy component <CardInfo />, inline styles (.content-cell-title and .content-cell-details) and delete ContentCell
-  let cmp: typeof SvelteComponent =
-    VOTING_UI === "legacy" ? CardInfo : ContentCell;
 </script>
 
 {#if visible}
-  <svelte:component this={cmp}>
+  <ProposalContentCell>
     <h4 slot="start">{$i18n.proposal_detail__ineligible.headline}</h4>
     <p class="description">{$i18n.proposal_detail__ineligible.text}</p>
     <ul>
@@ -40,7 +33,7 @@
         </li>
       {/each}
     </ul>
-  </svelte:component>
+  </ProposalContentCell>
 {/if}
 
 <style lang="scss">
