+++ frontend/src/lib/components/proposal-detail/MyVotes.svelte
@@ -1,7 +1,6 @@
 <script lang="ts">
   import type { ProposalInfo, NeuronId } from "@dfinity/nns";
   import { Vote } from "@dfinity/nns";
-  import CardInfo from "../ui/CardInfo.svelte";
   import { i18n } from "../../stores/i18n";
   import { IconThumbDown, IconThumbUp } from "@dfinity/gix-components";
   import { definedNeuronsStore } from "../../stores/neurons.store";
@@ -12,9 +11,7 @@
     type CompactNeuronInfo,
   } from "../../utils/neuron.utils";
   import Value from "../ui/Value.svelte";
-  import type { SvelteComponent } from "svelte";
-  import { VOTING_UI } from "../../constants/environment.constants";
-  import ContentCell from "../ui/ContentCell.svelte";
+  import ProposalContentCell from "./ProposalContentCell.svelte";
 
   export let proposalInfo: ProposalInfo;
 
@@ -44,21 +41,17 @@
       proposal: proposalInfo,
     });
   }
-
-  // TODO(L2-965): delete legacy component <CardInfo />, inline styles (.content-cell-title and .content-cell-details) and delete ContentCell
-  let cmp: typeof SvelteComponent =
-    VOTING_UI === "legacy" ? CardInfo : ContentCell;
 </script>
 
 {#if neuronsVotedForProposal.length}
-  <svelte:component this={cmp}>
+  <ProposalContentCell>
     <h4 slot="start">{$i18n.proposal_detail.my_votes}</h4>
     <ul>
       {#each neuronsVotedForProposal as neuron}
         <li
-          data-tid="neuron-data"
-          aria-label={voteMapper({ neuron: neuron.id, vote: neuron.vote })}
-          title={voteMapper({ neuron: neuron.id, vote: neuron.vote })}
+                data-tid="neuron-data"
+                aria-label={voteMapper({ neuron: neuron.id, vote: neuron.vote })}
+                title={voteMapper({ neuron: neuron.id, vote: neuron.vote })}
         >
           <p class="value">{neuron.id}</p>
           <p class="vote-details">
@@ -70,7 +63,7 @@
         </li>
       {/each}
     </ul>
-  </svelte:component>
+  </ProposalContentCell>
 {/if}
 
 <style lang="scss">
