+++ frontend/src/lib/components/proposal-detail/Proposal.svelte
