+++ frontend/src/lib/components/proposal-detail/ProposalContentCell.svelte
@@ -7,7 +7,7 @@
 </div>
 
 <style lang="scss">
-  @use "@dfinity/gix-components/styles/mixins/media";
+  @use "../../../../node_modules/@dfinity/gix-components/styles/mixins/media";
 
   .content-cell-details:last-of-type {
     margin-bottom: 0;
