+++ frontend/src/lib/components/proposal-detail/VotingCard/VotingCard.svelte
@@ -6,10 +6,8 @@
   } from "@dfinity/nns";
 
   import { getContext, onDestroy } from "svelte";
-  import { i18n } from "../../../stores/i18n";
   import { definedNeuronsStore } from "../../../stores/neurons.store";
   import { votingNeuronSelectStore } from "../../../stores/proposals.store";
-  import CardInfo from "../../ui/CardInfo.svelte";
   import VotingConfirmationToolbar from "./VotingConfirmationToolbar.svelte";
   import VotingNeuronSelect from "./VotingNeuronSelect.svelte";
   import {
@@ -22,7 +20,6 @@
     type VoteRegistration,
   } from "../../../stores/vote-registration.store";
   import { registerVotes } from "../../../services/vote-registration.services";
-  import { VOTING_UI } from "../../../constants/environment.constants";
   import { BottomSheet } from "@dfinity/gix-components";
 
   export let proposalInfo: ProposalInfo;
@@ -79,24 +76,9 @@
     unsubscribe();
     votingNeuronSelectStore.reset();
   });
-
-  // TODO(L2-965): delete legacy component <CardInfo />
 </script>
 
-{#if VOTING_UI === "legacy"}
-  {#if visible}
-    <CardInfo>
-      <h2 slot="start">{$i18n.proposal_detail__vote.headline}</h2>
-      <VotingNeuronSelect {proposalInfo} {voteRegistration} />
-      <VotingConfirmationToolbar
-        {proposalInfo}
-        {voteRegistration}
-        on:nnsConfirm={vote}
-        layout="legacy"
-      />
-    </CardInfo>
-  {/if}
-{:else if $definedNeuronsStore.length > 0}
+{#if $definedNeuronsStore.length > 0}
   <BottomSheet>
     {#if visible}
       <VotingConfirmationToolbar
