+++ frontend/src/lib/components/proposal-detail/VotingCard/VotingConfirmationToolbar.svelte
@@ -8,17 +8,14 @@
     mapProposalInfo,
     selectedNeuronsVotingPower,
   } from "../../../utils/proposals.utils";
-  import { replacePlaceholders } from "../../../utils/i18n.utils";
   import { busy } from "../../../stores/busy.store";
   import { Spinner } from "@dfinity/gix-components";
-  import { sanitize } from "../../../utils/html.utils";
   import type { VoteRegistration } from "../../../stores/vote-registration.store";
 
   const dispatch = createEventDispatcher();
 
   export let proposalInfo: ProposalInfo;
   export let voteRegistration: VoteRegistration | undefined = undefined;
-  export let layout: "legacy" | "modern";
 
   let id: ProposalId | undefined;
   let topic: string | undefined;
@@ -55,21 +52,9 @@
       voteType: selectedVoteType,
     });
   };
-
-  // TODO(L2-965): delete question
 </script>
 
-{#if layout === "legacy"}
-  <p class="question">
-    {@html replacePlaceholders($i18n.proposal_detail__vote.accept_or_reject, {
-      $id: `${id ?? ""}`,
-      $title: sanitize(title ?? ""),
-      $topic: sanitize(topic ?? ""),
-    })}
-  </p>
-{/if}
-
-<div role="toolbar" class={`${layout}`} data-tid="voting-confirmation-toolbar">
+<div role="toolbar" data-tid="voting-confirmation-toolbar">
   <button
     data-tid="vote-yes"
     {disabled}
@@ -109,21 +94,16 @@
   @use "@dfinity/gix-components/styles/mixins/media";
 
   [role="toolbar"] {
-    padding: var(--padding) 0 0;
-
     display: flex;
-    gap: var(--padding);
 
-    &.modern {
-      padding: var(--padding-2x) var(--padding-2x) 0;
-      justify-content: center;
-      gap: var(--padding-2x);
+    padding: var(--padding-2x) var(--padding-2x) 0;
+    justify-content: center;
+    gap: var(--padding-2x);
 
-      @include media.min-width(large) {
-        padding: 0;
-        justify-content: flex-start;
-        gap: var(--padding);
-      }
+    @include media.min-width(large) {
+      padding: 0;
+      justify-content: flex-start;
+      gap: var(--padding);
     }
   }
