+++ frontend/src/lib/components/proposals/ProposalCard.svelte
@@ -10,14 +10,10 @@
   import { routeStore } from "../../stores/route.store";
   import { AppPath } from "../../constants/routes.constants";
   import { mapProposalInfo } from "../../utils/proposals.utils";
-  import ProposalMeta from "./ProposalMeta.svelte";
   import type { Color } from "../../types/theme";
-  import Tag from "../ui/Tag.svelte";
 
   export let proposalInfo: ProposalInfo;
   export let hidden: boolean = false;
-  // TODO(L2-965): delete property and use modern
-  export let layout: "modern" | "legacy" = "legacy";
   import Value from "../ui/Value.svelte";
   import ProposalCountdown from "./ProposalCountdown.svelte";
 
@@ -41,56 +37,45 @@
 </script>
 
 <li class:hidden>
-  {#if layout === "legacy"}
-    <Card role="link" on:click={showProposal} testId="proposal-card">
-      <div slot="start" class="title-container">
-        <p class="title" {title}>{title}</p>
-      </div>
-      <Tag slot="end" {color}>{$i18n.status[ProposalStatus[status]] ?? ""}</Tag>
+  <Card
+    role="link"
+    on:click={showProposal}
+    testId="proposal-card"
+    withArrow={true}
+  >
+    <div class="card-meta id">
+      <Value ariaLabel={$i18n.proposal_detail.id_prefix}>{id}</Value>
+    </div>
 
-      <ProposalMeta {proposalInfo} showTopic />
-    </Card>
-  {:else}
-    <Card
-      role="link"
-      on:click={showProposal}
-      testId="proposal-card"
-      withArrow={true}
-    >
-      <div class="card-meta id">
-        <Value ariaLabel={$i18n.proposal_detail.id_prefix}>{id}</Value>
-      </div>
+    <div class="card-meta">
+      <span>{$i18n.proposal_detail.type_prefix}</span>
+      <Value ariaLabel={$i18n.proposal_detail.type_prefix}>{type}</Value>
+    </div>
 
+    <div class="card-meta">
+      <span>{$i18n.proposal_detail.topic_prefix}</span>
+      <Value>{topic ?? ""}</Value>
+    </div>
+
+    {#if proposer !== undefined}
       <div class="card-meta">
-        <span>{$i18n.proposal_detail.type_prefix}</span>
-        <Value ariaLabel={$i18n.proposal_detail.type_prefix}>{type}</Value>
+        <span>{$i18n.proposal_detail.proposer_prefix}</span>
+        <Value>{proposer}</Value>
       </div>
+    {/if}
 
-      <div class="card-meta">
-        <span>{$i18n.proposal_detail.topic_prefix}</span>
-        <Value>{topic ?? ""}</Value>
-      </div>
+    <blockquote class="title-placeholder">
+      <p class="description">{title}</p>
+    </blockquote>
 
-      {#if proposer !== undefined}
-        <div class="card-meta">
-          <span>{$i18n.proposal_detail.proposer_prefix}</span>
-          <Value>{proposer}</Value>
-        </div>
-      {/if}
+    <div class="card-meta">
+      <p class={`${color ?? ""} status`}>
+        {$i18n.status[ProposalStatus[status]] ?? ""}
+      </p>
 
-      <blockquote class="title-placeholder">
-        <p class="description">{title}</p>
-      </blockquote>
-
-      <div class="card-meta">
-        <p class={`${color ?? ""} status`}>
-          {$i18n.status[ProposalStatus[status]] ?? ""}
-        </p>
-
-        <ProposalCountdown {proposalInfo} />
-      </div>
-    </Card>
-  {/if}
+      <ProposalCountdown {proposalInfo} />
+    </div>
+  </Card>
 </li>
 
 <style lang="scss">
