+++ frontend/src/lib/components/proposals/ProposalsFilters.svelte
@@ -7,7 +7,6 @@
   import { proposalsFiltersStore } from "../../stores/proposals.store";
   import { enumsExclude } from "../../utils/enum.utils";
   import FiltersButton from "../ui/FiltersButton.svelte";
-  import { VOTING_UI } from "../../constants/environment.constants";
 
   let modalFilters: ProposalsFilterModalProps | undefined = undefined;
 
@@ -37,7 +36,7 @@
   }).length;
 </script>
 
-<div class={`filters ${VOTING_UI}`}>
+<div class="filters">
   <FiltersButton
     testId="filters-by-topics"
     totalFilters={totalFiltersTopic}
@@ -94,11 +93,6 @@
     display: flex;
     flex-wrap: wrap;
     padding: 0 0 var(--padding-3x);
-
-    // TODO(L2-965): delete legacy style
-    &.legacy {
-      padding: var(--padding) 0;
-    }
 
     --select-flex-direction: row-reverse;
