+++ frontend/src/lib/constants/environment.constants.ts
@@ -4,11 +4,8 @@ export const ROLLUP_WATCH: boolean = process.env.ROLLUP_WATCH === "true";
 export const FETCH_ROOT_KEY: boolean = process.env.FETCH_ROOT_KEY === "true";
 export const WASM_CANISTER_ID: string = String(process.env.WASM_CANISTER_ID);
 
-export const {
-  ENABLE_SNS,
-  VOTING_UI,
-}: { ENABLE_SNS: boolean; VOTING_UI: "legacy" | "modern" } = JSON.parse(
-  process.env.FEATURE_FLAGS ?? '{"ENABLE_SNS":false,"VOTING_UI":"legacy"}'
+export const { ENABLE_SNS }: { ENABLE_SNS: boolean } = JSON.parse(
+  process.env.FEATURE_FLAGS ?? '{"ENABLE_SNS":false}'
 );
 
 export const IS_TESTNET: boolean =
