+++ frontend/src/lib/i18n/en.json
@@ -423,7 +423,6 @@
     "confirm_reject_headline": "Reject Proposal",
     "confirm_reject_text": "You are about to cast $votingPower votes against this proposal, are you sure you want to proceed?",
     "vote_status": "Neuron $neuronId has voted $vote",
-    "accept_or_reject": "Do you want to adopt or reject Proposal <strong>$id</strong> &ndash; <strong>$title</strong> for <strong>$topic</strong>?",
     "cast_vote_neuronId": "Neuron ID $neuronId",
     "cast_vote_votingPower": "Voting power of $votingPower",
     "vote_adopt_in_progress": "Adopting proposal $topic ($proposalId). $status",
