+++ frontend/src/tests/lib/components/proposal-detail/VotingCard/VotingConfirmationToolbar.spec.ts
@@ -2,16 +2,13 @@
  * @jest-environment jsdom
  */
 
-import { Topic, Vote } from "@dfinity/nns";
-import type { Proposal } from "@dfinity/nns/dist/types/types/governance_converters";
+import { Vote } from "@dfinity/nns";
 import { fireEvent } from "@testing-library/dom";
 import { render, waitFor } from "@testing-library/svelte";
 import VotingConfirmationToolbar from "../../../../../lib/components/proposal-detail/VotingCard/VotingConfirmationToolbar.svelte";
 import { E8S_PER_ICP } from "../../../../../lib/constants/icp.constants";
 import { votingNeuronSelectStore } from "../../../../../lib/stores/proposals.store";
-import { replacePlaceholders } from "../../../../../lib/utils/i18n.utils";
 import { formatVotingPower } from "../../../../../lib/utils/neuron.utils";
-import en from "../../../../mocks/i18n.mock";
 import { mockNeuron } from "../../../../mocks/neurons.mock";
 import {
   mockProposalInfo,
@@ -27,8 +24,7 @@ describe("VotingConfirmationToolbar", () => {
   };
 
   const props = {
-    proposalInfo: mockProposalInfo,
-    layout: "modern",
+    proposalInfo: mockProposalInfo
   };
 
   beforeEach(() => {
@@ -139,33 +135,5 @@ describe("VotingConfirmationToolbar", () => {
     );
     expect(onConfirm).toBeCalled();
     expect(calledVoteType).toBe(Vote.No);
-  });
-
-  it("should display a question that repeats id and topic", async () => {
-    const { container } = render(VotingConfirmationToolbar, {
-      props: {
-        ...props,
-        layout: "legacy",
-      },
-    });
-
-    const testLabel = replacePlaceholders(
-      en.proposal_detail__vote.accept_or_reject,
-      {
-        $id: `${mockProposalInfo.id}`,
-        $title: `${(mockProposalInfo.proposal as Proposal).title}`,
-        $topic: en.topics[Topic[mockProposalInfo.topic]],
-      }
-    )
-      .replace(/<strong>/g, "")
-      .replace(/<\/strong>/g, "")
-      .replace("&ndash;", "–");
-
-    await waitFor(() => {
-      const { textContent }: HTMLParagraphElement = container.querySelector(
-        ".question"
-      ) as HTMLParagraphElement;
-      expect(textContent).toEqual(testLabel);
-    });
   });
 });
